+++ src/definition/functions.ts
@@ -1,3 +1,0 @@
-export const FUNCTIONS: { [FunctionName: string]: string } = {
-  state: "light",
-};